Checkpoint commit: refactoring the request code in TIDS for better reuse
  * Move tr_gss.[ch] to tr_gss_names.[ch], that is what the files contain
  * Add new tr_gss.[ch] containing generalized GSS request/response code
  * Refactor tids request handlers to use generalized code

  * First steps towards a monitoring interface handler, not functional
  * Rename listen_on_all_addrs() to tr_sock_listen_all()
  * Make better use of talloc in a few places
  * Clean up a few missing or unused #includes
  * Fix a few data types for the sake of pedantry

+++ include/tid_internal.h
@@ -97,7 +97,7 @@ struct tids_instance {
   TIDS_REQ_FUNC *req_handler;
   tids_auth_func *auth_handler;
   void *cookie;
-  uint16_t tids_port;
+  unsigned int tids_port;
   TR_NAME *gss_name;		/* GSS name client used for authentication */
 };
